src/requirements_server.txt
# SentinelCore — Central Server Dependencies
# Install on the machine running Kafka + PostgreSQL
#
# Usage (WSL/Linux):
#   pip install -r requirements_server.txt
#
# Components running on the server:
#   kafka_to_postgres.py  — Kafka consumer → PostgreSQL writer
#   api_server.py         — FastAPI dashboard backend
#   feature_builder.py    — ML snapshot generator
#   ml_engine.py          — Predictive Analytics module
#   analyzer.py           — offline event analyzer (optional)

# ── Kafka consumer ────────────────────────────────────────────────────────────
kafka-python-ng>=2.2.3

# ── PostgreSQL driver ─────────────────────────────────────────────────────────
psycopg2-binary>=2.9.11

# ── FastAPI + ASGI server ─────────────────────────────────────────────────────
fastapi>=0.135.1
uvicorn[standard]>=0.41.0
pydantic>=2.12.5

# ── Utilities (also needed by analyzer) ──────────────────────────────────────
requests>=2.32.5
psutil>=7.2.2

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=9.0.2
pytest-timeout>=2.3.1

# ── Machine Learning ──────────────────────────────────────────────────────────
scikit-learn>=1.5.0
numpy>=1.26.0
pandas>=2.2.0